@@ -1,17 +1,32 @@
 module github.com/decred/gominer
 
+go 1.18
+
 require (
 	github.com/barnex/cuda5 v0.0.0-20171012184954-da30a9b287d8
 	github.com/davecgh/go-spew v1.1.1
-	github.com/decred/dcrd/blockchain/standalone v1.0.0
-	github.com/decred/dcrd/chaincfg/chainhash v1.0.2
-	github.com/decred/dcrd/chaincfg/v3 v3.0.0
-	github.com/decred/dcrd/dcrutil/v3 v3.0.0
-	github.com/decred/dcrd/wire v1.4.0
+	github.com/decred/dcrd/blockchain/standalone/v2 v2.2.0
+	github.com/decred/dcrd/chaincfg/chainhash v1.0.4
+	github.com/decred/dcrd/chaincfg/v3 v3.2.0
+	github.com/decred/dcrd/dcrutil/v4 v4.0.1
+	github.com/decred/dcrd/wire v1.6.0
 	github.com/decred/go-socks v1.1.0
-	github.com/decred/slog v1.0.0
-	github.com/jessevdk/go-flags v1.4.0
+	github.com/decred/slog v1.2.0
+	github.com/jessevdk/go-flags v1.5.0
 	github.com/jrick/logrotate v1.0.0
 )
 
-go 1.12
+require (
+	github.com/agl/ed25519 v0.0.0-20170116200512-5312a6153412 // indirect
+	github.com/dchest/siphash v1.2.3 // indirect
+	github.com/decred/base58 v1.0.5 // indirect
+	github.com/decred/dcrd/crypto/blake256 v1.0.1 // indirect
+	github.com/decred/dcrd/crypto/ripemd160 v1.0.2 // indirect
+	github.com/decred/dcrd/dcrec v1.0.1 // indirect
+	github.com/decred/dcrd/dcrec/edwards/v2 v2.0.3 // indirect
+	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.2.0 // indirect
+	github.com/decred/dcrd/txscript/v4 v4.1.0 // indirect
+	github.com/klauspost/cpuid/v2 v2.0.9 // indirect
+	golang.org/x/sys v0.0.0-20210320140829-1e4c9ba3b0c4 // indirect
+	lukechampine.com/blake3 v1.2.1 // indirect
+)
