@@ -5,9 +5,9 @@ package cl
 */
 import "C"
 
-///////////////////////////////////////////////
-//OpenCL 1.1
-///////////////////////////////////////////////
+// /////////////////////////////////////////////
+// OpenCL 1.1
+// /////////////////////////////////////////////
 func CLUnloadCompiler() CL_int {
 	return CL_int(C.clUnloadCompiler())
 }
