@@ -8,7 +8,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/decred/dcrd/blockchain/standalone"
+	"github.com/decred/dcrd/blockchain/standalone/v2"
 	"github.com/decred/dcrd/chaincfg/chainhash"
 	"github.com/decred/dcrd/chaincfg/v3"
 
@@ -18,7 +18,6 @@ import (
 )
 
 var chainParams = chaincfg.MainNetParams()
-var deviceLibraryInitialized = false
 
 // Constants for fan and temperature bits
 const (
